backend/kanban/requirements.txt
django
djangorestframework
environs
django-cors-headers
psycopg2-binary
gunicorn
dj-database-url